+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,6 +1,18 @@
 # @ballerine/backoffice-v2
 
-## 0.5.48-11210907c.0
+## 0.5.49
+
+### Patch Changes
+
+- Ver bump
+- Updated dependencies
+  - @ballerine/blocks@0.1.27
+  - @ballerine/common@0.7.43
+  - @ballerine/ui@0.3.29
+  - @ballerine/workflow-browser-sdk@0.5.43
+  - @ballerine/workflow-node-sdk@0.5.43
+
+## 0.5.48
 
 ### Patch Changes
 
@@ -10,6 +22,9 @@
   - @ballerine/ui@0.3.29-11210907c.0
   - @ballerine/workflow-browser-sdk@0.5.42-11210907c.0
   - @ballerine/workflow-node-sdk@0.5.42-11210907c.0
+  - @ballerine/common@0.7.42
+  - @ballerine/workflow-browser-sdk@0.5.42
+  - @ballerine/workflow-node-sdk@0.5.42
 
 ## 0.5.47