+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,10 +1,18 @@
 # @ballerine/workflow-node-sdk
 
-## 0.5.42-11210907c.0
+## 0.5.43
 
 ### Patch Changes
 
-- @ballerine/workflow-core@0.5.42-11210907c.0
+- Ver bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.43
+
+## 0.5.42
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.42
 
 ## 0.5.41