+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/websocket-service
 
+## 0.0.18
+
+### Patch Changes
+
+- Ver bump
+
 ## 0.0.17
 
 ### Patch Changes